Added risk score logic to load account
