+++ runtime/src/bank.rs
@@ -4555,38 +4555,6 @@ impl Bank {
         let sanitized_txs = batch.sanitized_transactions();
 
         // println!("processing transactions: {}", sanitized_txs.len());
-        let tag_expected = "AI_SCORE";
-
-        for tx in sanitized_txs {
-            for instruction in tx.message().instructions().iter() {
-                let program_id = tx.message().account_keys()[instruction.program_id_index as usize];
-        
-                // Check if the instruction is from the memo program
-                if program_id == Pubkey::from_str("meB2UQZBLPvWz3xmahLzWh6E3fLoPRKVwGztEHd933X").unwrap() {
-                    if let Ok(memo_str) = std::str::from_utf8(&instruction.data) {
-                        // Parse the string as JSON
-                        if let Ok(memo_json) = serde_json::from_str::<serde_json::Value>(memo_str) {
-                            if memo_json["tag"].as_str() == Some(tag_expected) {
-                                let version = memo_json["version"].as_u64().unwrap_or(0);
-                                let ai_score = memo_json["ai_score"].as_f64().unwrap_or(0.0) as f32;
-                                let wallet_address = memo_json["wallet_address"].as_str().unwrap_or("").to_string();
-        
-                                if let Some(payer_account_info) = tx.message().account_keys().get(0) {
-                                    let wallet = payer_account_info.to_string();
-                                    let reward_account = wallet.clone();
-                                    let risk_score = 5.0;
-                                    let timeout = 6000;
-        
-                                    // Update the risk scores
-                                    println!("Found AI score with version {}: {:?}, Wallet: {}", version, ai_score, &wallet_address);
-        
-                                    update_risk_scores(wallet_address, reward_account, risk_score, timeout);
-                                }
-                            }
-                        }
-                    }}}
-        }
-        
 
         let mut error_counters = TransactionErrorMetrics::default();
